requirements.txt
django==1.10.5
djangorestframework==3.5.4
django-oauth-toolkit==0.12.0
psycopg2==2.6.2
celery==4.0.2
redis==2.10.5